# [libc++] Fix sub-overflow in std::gcd implementation