Fix #117249