@serge-sans-paille 2024-11-30
+1 for the regression test. Considering the algorithm, it's based on the same idea, just a "different wording". I'll run the benchmark .