@serge-sans-paille 2024-12-05
after:

```
# | ---------------------------------------------------------
# | Benchmark               Time             CPU   Iterations
# | ---------------------------------------------------------
# | bm_gcd_random      101356 ns       101209 ns         7000
# | bm_gcd_trivial       1.85 ns         1.85 ns    378997661
# | bm_gcd_complex       34.2 ns         34.2 ns     20482199
```

before:

```
# | ---------------------------------------------------------
# | Benchmark               Time             CPU   Iterations
# | ---------------------------------------------------------
# | bm_gcd_random      105106 ns       104955 ns         7000
# | bm_gcd_trivial       1.85 ns         1.85 ns    378980115
# | bm_gcd_complex       39.8 ns         39.8 ns     17596552
```